// client/src/services/workout/types.ts
export interface WorkoutProps {
    _id: string;
    date: string;
    description: string;
    createdAt?: string;
    updatedAt?: string;
    isPrivate: boolean;
    name: string;
    exercises: WorkoutExerciseProps[];
    programUid?: string;
    status: WorkoutStatus;
    strainRating?: number;
    reflection?: string;
    likeUids?: string[];
    imageUri?: string;
    localImageUri?: string;
    imageBase64?: string;
    imageId?: string;
    userUid?: string;
    healthData?: HealthDataProps;
    type: HealthActivity;
}
// ...
export interface HealthDataProps {
    _id?: string;
    id?: string;
    workoutUid?: string,
    activityId: string,
    activityName: string,
    sourceName: string,
    duration: number,
    calories: number,
    distance: number,
    heartRates?: number[],
    date: string,
    end?: string,
    start?: string,
    disMeas?: HealthDisMeas;
}

export enum HealthDisMeas {
    mi = 'mi',
    km = 'km',
    m = 'm',
    yd = 'yd'
}


export interface WorkoutExercisesObjProps {
    [group: string]: WorkoutExerciseProps[];
}

export enum WorkoutStatus {
    pending = 'pending',
    inProgress = 'inProgress',
    completed = 'completed'
}

export interface WorkoutExerciseProps {
    _id?: string;
    remove?: boolean;
    workoutUid?: string;
    tempId?: string;
    data: WorkoutExerciseDataProps[];
    exerciseUid?: string;
    exercise?: ExerciseProps;
    group: number;
    order: number;
    sets?: number;
    reps?: number;
    calcRef?: number;
    comment?: string;
    date?: string;
    programWorkoutUid?: string;
    programTemplateUid?: string
}

export interface WorkoutExerciseDataProps {
    _id?: string;
    reps: number;
    performVal?: number;
    predictVal: number;
    pct?: number;
    completed?: boolean;
    warmup?: boolean;
}